Refactor `Commands` into relevant test packages
