+++ src/test/java/trackr/logic/commands/order/AddOrderCommandIntegrationTest.java
@@ -1,3 +1,4 @@
+package trackr.logic.commands.order;
 //package trackr.logic.commands;
 //
 //import static trackr.logic.commands.CommandTestUtil.assertCommandFailure;
