+++ src/test/java/trackr/logic/commands/supplier/ClearSupplierCommandTest.java
@@ -1,4 +1,4 @@
-package trackr.logic.commands;
+package trackr.logic.commands.supplier;
 
 import static trackr.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static trackr.testutil.TypicalMenuItems.getTypicalMenu;
@@ -8,7 +8,6 @@ import static trackr.testutil.TypicalTasks.getTypicalTaskList;
 
 import org.junit.jupiter.api.Test;
 
-import trackr.logic.commands.supplier.ClearSupplierCommand;
 import trackr.logic.parser.exceptions.ParseException;
 import trackr.model.Model;
 import trackr.model.ModelEnum;
