+++ src/test/java/trackr/logic/commands/supplier/EditSupplierCommandTest.java
@@ -1,4 +1,4 @@
-package trackr.logic.commands;
+package trackr.logic.commands.supplier;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -21,8 +21,6 @@ import org.junit.jupiter.api.Test;
 
 import trackr.commons.core.Messages;
 import trackr.commons.core.index.Index;
-import trackr.logic.commands.supplier.ClearSupplierCommand;
-import trackr.logic.commands.supplier.EditSupplierCommand;
 import trackr.logic.parser.exceptions.ParseException;
 import trackr.model.Menu;
 import trackr.model.Model;
