+++ src/test/java/trackr/logic/commands/supplier/FindSupplierCommandTest.java
@@ -1,4 +1,4 @@
-package trackr.logic.commands;
+package trackr.logic.commands.supplier;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -18,7 +18,6 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
 
-import trackr.logic.commands.supplier.FindSupplierCommand;
 import trackr.logic.parser.exceptions.ParseException;
 import trackr.model.Model;
 import trackr.model.ModelEnum;
