+++ src/test/java/trackr/logic/commands/task/AddTaskCommandTest.java
@@ -1,4 +1,4 @@
-package trackr.logic.commands;
+package trackr.logic.commands.task;
 
 import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -11,8 +11,8 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
+import trackr.logic.commands.CommandResult;
 import trackr.logic.commands.exceptions.CommandException;
-import trackr.logic.commands.task.AddTaskCommand;
 import trackr.model.ModelEnum;
 import trackr.model.ReadOnlyTaskList;
 import trackr.model.TaskList;
