+++ src/test/java/trackr/logic/commands/task/ClearTaskCommandTest.java
@@ -1,4 +1,4 @@
-package trackr.logic.commands;
+package trackr.logic.commands.task;
 
 import static trackr.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static trackr.testutil.TypicalMenuItems.getTypicalMenu;
@@ -8,37 +8,36 @@ import static trackr.testutil.TypicalTasks.getTypicalTaskList;
 
 import org.junit.jupiter.api.Test;
 
-import trackr.logic.commands.order.ClearOrderCommand;
 import trackr.logic.parser.exceptions.ParseException;
 import trackr.model.Model;
 import trackr.model.ModelEnum;
 import trackr.model.ModelManager;
 import trackr.model.UserPrefs;
 
-public class ClearOrderCommandTest {
+public class ClearTaskCommandTest {
 
     @Test
-    public void execute_emptyOrderList_success() throws ParseException {
+    public void execute_emptyTaskList_success() throws ParseException {
         Model model = new ModelManager();
         Model expectedModel = new ModelManager();
 
-        assertCommandSuccess(new ClearOrderCommand(),
+        assertCommandSuccess(new ClearTaskCommand(),
                 model,
-                String.format(ClearOrderCommand.MESSAGE_SUCCESS, ModelEnum.ORDER),
+                String.format(ClearTaskCommand.MESSAGE_SUCCESS, ModelEnum.TASK),
                 expectedModel);
     }
 
     @Test
-    public void execute_nonEmptyOrderList_success() throws ParseException {
+    public void execute_nonEmptyTaskList_success() throws ParseException {
         Model model = new ModelManager(getTypicalSupplierList(), getTypicalTaskList(),
                 getTypicalMenu(), getTypicalOrderList(), new UserPrefs());
         Model expectedModel = new ModelManager(getTypicalSupplierList(), getTypicalTaskList(),
                 getTypicalMenu(), getTypicalOrderList(), new UserPrefs());
-        expectedModel.setItemList(ModelEnum.ORDER);
+        expectedModel.setItemList(ModelEnum.TASK);
 
-        assertCommandSuccess(new ClearOrderCommand(),
+        assertCommandSuccess(new ClearTaskCommand(),
                 model,
-                String.format(ClearOrderCommand.MESSAGE_SUCCESS, ModelEnum.ORDER),
+                String.format(ClearTaskCommand.MESSAGE_SUCCESS, ModelEnum.TASK),
                 expectedModel);
     }
