+++ src/test/java/trackr/logic/commands/task/ListTaskCommandTest.java
@@ -1,7 +1,7 @@
-package trackr.logic.commands;
+package trackr.logic.commands.task;
 
 import static trackr.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static trackr.logic.commands.CommandTestUtil.showSupplierAtIndex;
+import static trackr.logic.commands.CommandTestUtil.showTaskAtIndex;
 import static trackr.testutil.TypicalIndexes.INDEX_FIRST_OBJECT;
 import static trackr.testutil.TypicalMenuItems.getTypicalMenu;
 import static trackr.testutil.TypicalOrders.getTypicalOrderList;
@@ -11,7 +11,7 @@ import static trackr.testutil.TypicalTasks.getTypicalTaskList;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import trackr.logic.commands.supplier.ListSupplierCommand;
+import trackr.logic.commands.ListItemCommand;
 import trackr.logic.parser.exceptions.ParseException;
 import trackr.model.Model;
 import trackr.model.ModelEnum;
@@ -19,9 +19,9 @@ import trackr.model.ModelManager;
 import trackr.model.UserPrefs;
 
 /**
- * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
+ * Contains integration tests (interaction with the Model) and unit tests for ListTaskCommand.
  */
-public class ListSupplierCommandTest {
+public class ListTaskCommandTest {
 
     private Model model;
     private Model expectedModel;
@@ -35,19 +35,19 @@ public class ListSupplierCommandTest {
     }
 
     @Test
-    public void execute_listIsNotFiltered_showsSameList() throws ParseException {
-        assertCommandSuccess(new ListSupplierCommand(),
+    public void execute_taskListIsNotFiltered_showsSameList() throws ParseException {
+        assertCommandSuccess(new ListTaskCommand(),
                 model,
-                String.format(ListItemCommand.MESSAGE_SUCCESS, ModelEnum.SUPPLIER.toString().toLowerCase()),
+                String.format(ListItemCommand.MESSAGE_SUCCESS, ModelEnum.TASK.toString().toLowerCase()),
                 expectedModel);
     }
 
     @Test
-    public void execute_listIsFiltered_showsEverything() throws ParseException {
-        showSupplierAtIndex(model, INDEX_FIRST_OBJECT);
-        assertCommandSuccess(new ListSupplierCommand(),
+    public void execute_taskListIsFiltered_showsEverything() throws ParseException {
+        showTaskAtIndex(model, INDEX_FIRST_OBJECT);
+        assertCommandSuccess(new ListTaskCommand(),
                 model,
-                String.format(ListItemCommand.MESSAGE_SUCCESS, ModelEnum.SUPPLIER.toString().toLowerCase()),
+                String.format(ListItemCommand.MESSAGE_SUCCESS, ModelEnum.TASK.toString().toLowerCase()),
                 expectedModel);
     }
 }
